backend/requirements.txt
fastapi
uvicorn
pyjwt
pytest